+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Sat May 31 13:41:24 PDT 2014"
+buildTime = "Sat May 31 14:54:49 PDT 2014"