+++ CHANGES.md
@@ -18,6 +18,13 @@
         sBranch is used. Of course, if time-out happens, SBV will
 	assume the branch is feasible, in which case symbolic-termination
 	may come back to bite you.)
+ * New functionality:
+      * Add predicate 'isSNaN' which allows testing 'SFloat'/'SDouble' values
+        for nan-ness. This is similar to the Prelude function 'isNaN', except
+	the Prelude version requires a RealFrac instance, which unfortunately is
+	not currently implementable for cases. (Requires trigonometric functions etc.)
+	Thus, we provide 'isSNaN' separately (along with the already existing
+	'isFPPoint') to simplify reasoning with floating-point.
  * Examples:
      * Add Data/SBV/Examples/Misc/SBranch.hs, to illustrate the use of sBranch.
  * Bug fixes: